chore(tests): split up e2e tests - tfm amendments
