+++ .github/workflows/test.yml
@@ -587,14 +587,7 @@ jobs:
             'case-amendments/show-amendment-button/MIN/*.spec.js',
             'case-amendments/tasks/*.spec.js',
             'case-amendments/underwriting-page-unassigned/*.spec.js',
-            'case-amendments/user-flow/step-1/*.spec.js',
-            'case-amendments/user-flow/step-2/*.spec.js',
-            'case-amendments/user-flow/step-3/*.spec.js',
-            'case-amendments/user-flow/step-4/*.spec.js',
-            'case-amendments/user-flow/step-5/*.spec.js',
-            'case-amendments/user-flow/step-6/*.spec.js',
-            'case-amendments/user-flow/step-7/*.spec.js',
-            'case-amendments/user-flow/step-8/*.spec.js',
+            'case-amendments/user-flow/*.spec.js',
             'case-underwriting/**/*.spec.js',
             'deals/**/*.spec.js',
             'facility/**/*.spec.js',
