+++ e2e-tests/tfm/cypress/e2e/journeys/case-amendments/user-flow/1. add-amendment-request.spec.js
@@ -1,10 +1,10 @@
-import relative from '../../../../relativeURL';
-import { errorSummary } from '../../../../partials';
-import facilityPage from '../../../../pages/facilityPage';
-import amendmentsPage from '../../../../pages/amendments/amendmentsPage';
-import MOCK_DEAL_AIN from '../../../../../fixtures/deal-AIN';
-import dateConstants from '../../../../../../../e2e-fixtures/dateConstants';
-import { PIM_USER_1, BANK1_MAKER1, ADMIN } from '../../../../../../../e2e-fixtures';
+import relative from '../../../relativeURL';
+import { errorSummary } from '../../../partials';
+import facilityPage from '../../../pages/facilityPage';
+import amendmentsPage from '../../../pages/amendments/amendmentsPage';
+import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
+import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
+import { PIM_USER_1, BANK1_MAKER1, ADMIN } from '../../../../../../e2e-fixtures';
 
 context('Amendments - Request date', () => {
   let dealId;
